Implement ConditionalStatement trait for select, delete, update
Note that this changes the semantics of `and_where` and `or_where` for
Delete and Update, to align them with Select. The difference is in the
operator precedence.

Previously, we had:

```rust
Query::delete()
  .from_table(Glyph::Table)
  .or_where(Expr::col(Glyph::Id).eq(1))
  .or_where(Expr::col(Glyph::Id).eq(2))
  .and_where(Expr::col(Glyph::Id).eq(3))
  .to_string(PostgresQueryBuilder);
```

equal to

```sql
DELETE FROM "glyph" WHERE (("id" = 1) OR ("id" = 2)) AND ("id" = 3)
```

I.e. it was using left-associativity. With this change, we now have the
less surprising:

```sql
DELETE FROM "glyph" WHERE "id" = 1 OR "id" = 2 AND "id" = 3
```

I.e. using the natural operator precedence from SQL, with AND before OR.
In that case, the parentheses would be:

```sql
DELETE FROM "glyph" WHERE "id" = 1 OR ("id" = 2 AND "id" = 3)
```