Fix serializing iden as a value in `ALTER TYPE ... RENAME TO ...` statements  (#924)
<!--

Thank you for contributing to this project!

If you need any help please feel free to contact us on Discord:
https://discord.com/invite/uCPdDXzbdv
Or, mention our core members by typing `@GitHub_Handle` on any issue /
PR

Add some test cases! It help reviewers to understand the behaviour and
prevent it to be broken in the future.

-->

## PR Info

<!-- mention the related issue -->
- Closes #898

<!-- is this PR depends on other PR? (if applicable) -->
- Dependencies:
  - <!-- PR link -->

<!-- any PR depends on this PR? (if applicable) -->
- Dependents:
  - <!-- PR link -->

## New Features

- [ ] <!-- what are the new features? -->

## Bug Fixes

- [x] Fix serializing iden as a value in `ALTER TYPE ... RENAME TO ...`
statements

## Breaking Changes

- [ ] <!-- any change in behaviour or method signature? is it backward
compatible? -->

## Changes

- [ ] <!-- any other non-breaking changes to the codebase -->