Support jiff (#941)
<!--

Thank you for contributing to this project!

If you need any help please feel free to contact us on Discord:
https://discord.com/invite/uCPdDXzbdv
Or, mention our core members by typing `@GitHub_Handle` on any issue /
PR

Add some test cases! It help reviewers to understand the behaviour and
prevent it to be broken in the future.

-->

## PR Info

<!-- mention the related issue -->
- Closes #851

<!-- is this PR depends on other PR? (if applicable) -->
- Dependencies:
  - <!-- PR link -->

<!-- any PR depends on this PR? (if applicable) -->
- Dependents:
  - <!-- PR link -->

## New Features

- [ ] <!-- what are the new features? -->

## Bug Fixes

- [ ] <!-- if it fixes a bug, please provide a brief analysis of the
original bug -->

## Breaking Changes

- [ ] <!-- any change in behaviour or method signature? is it backward
compatible? -->

## Changes

- [ ] <!-- any other non-breaking changes to the codebase -->