Refactor conversions around qualified identifiers (#953)
There was some annoying repetition in this area. `TableName` and
`ColumnName` each had three copy-pasted `From` impls. Now they have one,
which will be easier to copy-paste when I add `TypeName` and
`FunctionName`. I came up with this while re-doing
https://github.com/SeaQL/sea-query/pull/922 to add `TypeName`.

Related to https://github.com/SeaQL/sea-query/pull/927.

`types.rs` is getting too long and starting to have visible disjoint
"submodules", so I started splitting it.

## PR Info

- Dependents:
  - https://github.com/SeaQL/sea-query/pull/922

## New Features

- [x] The new traits are technically an addition, but I wouldn't
advertise it. They are not meant to be used directly. I've documented
this on the traits.