Allow qualified type names in `cast_as_quoted` (#922)
As discussed here:
https://github.com/SeaQL/sea-query/issues/827#issuecomment-3125338008

## PR Info

- Closes #827

- Dependencies:
  - #921
  - #953

## Changes

The PR has been updated. The up-to-date description is in the changelog
in the diff.

---

**Original, outdated note:**

Merge after #921. Marking as draft until then. The first commit comes as
a dependency from that PR. If you want to review the diff before merging
that PR, review only the second commit.