Change the shared lock syntax to 'LOCK IN SHARE MODE' in MSQL for MariaDB support (#980)
<!--

Thank you for contributing to this project!

If you need any help please feel free to contact us on Discord:
https://discord.com/invite/uCPdDXzbdv
Or, mention our core members by typing `@GitHub_Handle` on any issue /
PR

Add some test cases! It help reviewers to understand the behaviour and
prevent it to be broken in the future.

-->

## PR Info

<!-- mention the related issue -->
- Closes https://github.com/SeaQL/sea-orm/issues/2107

## New Features

- [ ] None <!-- what are the new features? -->

## Bug Fixes

- [x] Fixes a syntax error when using `lock_shared()` on MariaDB. <!--
if it fixes a bug, please provide a brief analysis of the original bug
-->

## Breaking Changes

- [ ] None <!-- any change in behaviour or method signature? is it
backward compatible? -->

## Changes

- [x] Create a `lock_phrase` method in `QueryBuilder` so that it can be
overridden in `MySqlQueryBuilder` to use `LOCK IN SHARE MODE`.

---------

Co-authored-by: 狐狸 <134658521+Huliiiiii@users.noreply.github.com>