Re-exporting stronger-pinned third-party types (#1012)
<!--

Thank you for contributing to this project!

If you need any help please feel free to contact us on Discord:
https://discord.com/invite/uCPdDXzbdv
Or, mention our core members by typing `@GitHub_Handle` on any issue /
PR

Add some test cases! It help reviewers to understand the behaviour and
prevent it to be broken in the future.

-->

## PR Info

- Closes https://github.com/SeaQL/sea-query/issues/1011

## Changes

- [x] modified use -> pub use for some of the optional dependency types
in value.rs

---------

Co-authored-by: Georg Schulz-Allgaier <georg_schulz-allgaier@genua.de>