Implement Postgres ON CONFLICT ON CONSTRAINT (#1016)
## PR Info

- Closes: https://github.com/SeaQL/sea-query/issues/1013

## New Features

- [ ] Refactor OnConflict enum to conform the Postgres specification,
see the related issue.
- [ ] Possible change in API
- OnConflict::exprs() no longer appends to the list of columns, it has
`add_column` and `add_expr` function for that purpose.

## Breaking Changes

- [ ] Change OnConflict enum to enable the mix of columns and expression
list, or a single constraint name

```rust
impl OnConflict {
    pub fn column<C>(column: C) -> Self
    where
        C: IntoIden;

    pub fn columns<I, C>(columns: I) -> Self
    where
        C: IntoIden,
        I: IntoIterator<Item = C>;

    pub fn constraint(constraint: &str) -> Self;

    pub fn expr<T>(&mut self, expr: T) -> &mut Self
    where
        T: Into<Expr>;

    pub fn exprs<I, T>(&mut self, exprs: I) -> &mut Self
    where
        T: Into<Expr>,
        I: IntoIterator<Item = T>;
}
```
these functions replace the current data except the exprs which appens,
which is a bit inconsistent. Also with this API is not possible to make
a mix of column/expr. Is it a good idea to add like

```rust
    pub fn push_column<C>(&mut self, column: C);

    pub fn push_expr<T>(&mut self, expr: T);
```

To append to the existing targets (unsetting constraint probably since
they are mutually exclusive)? Also changing exprs to set targets not
append to them.

---------

Co-authored-by: Huliiiiii <134658521+Huliiiiii@users.noreply.github.com>
Co-authored-by: Chris Tsang <chris.2y3@outlook.com>